lambda/generate_website/requirements.txt
google-generativeai==0.3.2
langgraph==0.0.26
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
boto3==1.34.0
